@@ -1,11 +1,10 @@
 name: MexAppSwiftHttpWrapper
 description: A new flutter plugin project.
 version: 0.0.1
-author:
-homepage:
+homepage: https://aufwallet.com
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <4.0.0"
   flutter: ">=1.20.0 <2.0.0"
 
 dependencies:
