@@ -11,8 +11,8 @@ class MyApp extends StatefulWidget {
 }
 
 class _MyAppState extends State<MyApp> {
-  String _response;
-  SwiftHttpError _error;
+  String? _response;
+  SwiftHttpError? _error;
 
   @override
   void initState() {
@@ -23,15 +23,16 @@ class _MyAppState extends State<MyApp> {
   _getMarkets() async {
     await MexAppSwiftHttpWrapper.setTimeout(5);
     await MexAppSwiftHttpWrapper.setRetryCount(0);
-    final response = await MexAppSwiftHttpWrapper.request(NetworkRequest()
-      ..apiHost = 'https://app-api-int.duedex.com'
-      ..endpoint = '/v1/instrument'
-      ..method = 'get');
+    final response = await MexAppSwiftHttpWrapper.request(NetworkRequest(
+      apiHost: 'https://app-api-int.jfrex.com',
+      endpoint: '/v1/instrument',
+      method: 'get'
+    ));
 
     if (response is SwiftHttpError) {
       setState(() => _error = response);
     } else {
-      setState(() => _response = response);
+      setState(() => _response = response as String?);
     }
   }
 
@@ -54,9 +55,9 @@ class _MyAppState extends State<MyApp> {
                   child: CircularProgressIndicator(),
                 ),
               if (_response != null)
-                _text('Success response: ' + _response) ,
+                _text('Success response: ' + _response!) ,
               if (_error != null)
-                _text('Error: ' + _error.toJson().toString())
+                _text('Error: ' + _error!.toJson().toString())
             ],
           ),
         ),
